use `bw groups` for more group membership tests
